requirements.txt
bcrypt==4.1.2

# Special hardware imports (Raspberry Pi specific)
spidev==3.6  # Note: Version might vary based on system installation
smbus2==0.4.1  # More modern Python 3 compatible version
# Gamepad control requirements - pygame only version
pygame==2.6.1

#normal imports
annotated-types==0.7.0
anyio==4.8.0
bidict==0.23.1
click==8.1.8
fastapi==0.115.6
h11==0.16.0
idna==3.10
lgpio==0.2.2.0
mysql-connector-python==9.1.0
pydantic==2.10.5
pydantic_core==2.27.2
python-engineio==4.11.2
python-socketio==5.12.1
rpi-lgpio==0.6
simple-websocket==1.1.0
sniffio==1.3.1
starlette==0.41.3
typing_extensions==4.12.2
uvicorn==0.34.0
wsproto==1.2.0